deleting comments new exceptions in repo

+++ Repositories/CocktailRecipeRepository.cs
@@ -1,4 +1,5 @@
 ﻿using Drinks_app.Data;
+using Drinks_app.Exception;
 using Drinks_app.Models;
 using Drinks_app.Repositories.IRepositories;
 using Microsoft.EntityFrameworkCore;
@@ -22,29 +23,28 @@ namespace Drinks_app.Repositories
             _db.SaveChanges();
 
         }
-
-        //[DataObjectMethod(DataObjectMethodType.Delete)]
         public void DeleteCocktailRecipe(long id)
         {
-            //using (var db = _db.CocktailRecipes()) {
                 var deleteCocktailRecipe = (from recipe
                                            in _db.CocktailRecipes
                                            where recipe.Id == id
                                            select recipe).Include(recipe => recipe.User).Include(recipe => recipe.Ingredients).FirstOrDefault();
-                if (deleteCocktailRecipe != null)
+
+            if (deleteCocktailRecipe == null) throw new NotFoundException("Cocktail Recipe is not found");
+
+            if (deleteCocktailRecipe != null)
                 {
                 _db.CocktailRecipes.Remove(deleteCocktailRecipe);
               
                 _db.SaveChanges();
 
                 }
-            //}
         }
 
 
         public IEnumerable<CocktailRecipe> GetAllCocktailRecipe()
         {
-            var AllCocktailRecipes = from c in _db.CocktailRecipes
+            var allCocktailRecipes = from c in _db.CocktailRecipes
                                      select new CocktailRecipe
                                      {
                                          Id = c.Id,
@@ -52,7 +52,10 @@ namespace Drinks_app.Repositories
                                          Recipe = c.Recipe,
                                          User = c.User,
                                      };
-            return AllCocktailRecipes;
+
+            if (allCocktailRecipes == null) throw new NotFoundException("Cocktail Recipe is not found");
+
+            return allCocktailRecipes;
         }
 
             public CocktailRecipe GetCocktailRecipeById(long id)
@@ -62,6 +65,8 @@ namespace Drinks_app.Repositories
                                         in _db.CocktailRecipes
                                         where i.Id == id
                                         select i).FirstOrDefault();
+
+            if (getCocktailRecipeById == null) throw new NotFoundException("Cocktail Recipe is not found");
                                      
             return getCocktailRecipeById;
             
@@ -69,20 +74,20 @@ namespace Drinks_app.Repositories
         public void UpdateCocktailRecipe(CocktailRecipe cocktailRecipe)
         {
 
-            var UpdateCocktailRecipe = from updateRecipe
+            var updateCocktailRecipe = from updateRecipe
                                    in _db.CocktailRecipes
                                         where updateRecipe.Id == cocktailRecipe.Id
                                         select updateRecipe;
-            foreach(CocktailRecipe recip in UpdateCocktailRecipe)
+
+            if (updateCocktailRecipe == null) throw new NotFoundException("Cocktail Recipe is not found");
+
+            foreach (CocktailRecipe recip in updateCocktailRecipe)
             {
                 recip.Name = cocktailRecipe.Name;
                 recip.Recipe = cocktailRecipe.Recipe;
             }
             
-           
             _db.SaveChanges();
-            //UpdateCocktailRecipe.is_default = false;
-            //Context.SaveChanges();
         }
     } 
 }
